ForecastSchedule element name mismatch with XSD
In the XSD for CommonModule that was submitted to NAESB, the "irregularTimePoint" element is not plural.  After checking with Shawn Hu, this is the explanation for why:

"I believe the 's' was there in an early version since according to the CIM convention, an XSD element name goes with the UML association role name. And a role name usually goes with the multiplicity such as IrregularTimePoints for an 0..* relationship but not always. I checked the model again, the role name is not present in the current model hence its end class name is used as default"